requirements.txt
asgiref==3.3.4
boto3==1.17.72
botocore==1.20.72
Django==3.2.2
django-storages==1.11.1
ImageHash==4.2.0
jmespath==0.10.0
mysqlclient==2.0.3
numpy==1.20.2
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
s3transfer==0.4.2
scipy==1.6.3
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.4
